Add option to always show news bins in the reduced feed list
Add a new option "Always show in Reduced Feed List" to the news bin
properties so an user can make it appear there even when there are no
unread items. This enables some new and very practical workflows for
users wanting to use news bins to sort items or prepare reading lists
directly from the reduced view.

A new option always has a complexity cost for the user, but this one
seems worth it: the alternative of always showing news bins in the
reduced view, as happens with the search folders, can make the list
less useful for people who do not want all of them to be always there.

The setting is persisted in feedlist.opml, requires no changes to the
database schema and is forward and backwards compatible except, of
course, for the value being lost when loaded in a previews version.
Editing the option requires changes to the dialogs used to create news
bins and do not allow reusing the "rename node" dialog as a properties
dialog for this node type anymore. To prevent code duplication, the new
dialog is used both as a creating and a properties dialog.