Fix RCE vulnerability on feed auto-discovery
Liferea feed auto-discovery feature can create subscriptions to command
feeds, allowing malicious websites to run arbitrary commands on the
local system.

For example, trying feed discovery in a site which returns a document
like this:

    <!DOCTYPE html>
    <html>
     <head>
      <title>Feed auto-discovery RCE PoC</title>
      <link rel="alternate" type="application/rss+xml" href="|date &gt;/tmp/bad-feed-discovery.txt">
     </head>
     <body>
     Oooops.
     </body>
    </html>

will cause Liferea to subscribe to command
"date >/tmp/bad-feed-discovery.txt". The command runs before the user
has a chance of checking the subscription.

This happens because function common_build_url, used to normalize the
discovered URL candidates, fails to create an absolute URL if any of
the parts has some special characters (including "|"), returning the
original, non-escaped, relative URL instead. Inner working of the
failure is a bit tricky:

1. internally, common_build_url calls common_uri_escape to escape the
   URL, then xmlBuildURI to make an absolute URL from the returned
   values;

2. common_uri_escape is almost a wrapper for xmlURIEscape (it escapes
   spaces by itself). If the call to xmlURIEscape suceeds, it returns
   the result; if it fails, returns a copy of the original string. A
   "|" in the string will make xmlURIEscape fail, then common_uri_escape
   will give up and return a copy of the original string with that
   non-escaped "|";

3. then, common_build_url receives the returned string (with the
   non-escaped "|") and call xmlBuildURI to make an absolute URL. This
   also fails because of the "|", so common_build_url returns another
   copy of the source string with the non-escaped "|".

Immediate fix is to add a special substitution for "|" in
common_uri_escape (as already done with spaces); this will not work for
all cases. To ensure no command feeds are created from here, a new
validation is added to html_auto_discover_feed to ensure that only URLs
starting with "http://" or "https://" are accepted.  This has the side
effect of preventing auto-discovery from returning feeds that use other
protocol, but this not seems to be a problem.